+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 All Notable changes to `flipboxdigital\craft-sortable-associations` will be documented in this file
 
+## Unreleased
+### Changed
+- `SortableAssociation` record must implement an association service.
+
 ## 1.0.0-rc
 - Initial release!